+++ tests/retrostar/test_retrostar.py
@@ -1,3 +1,5 @@
+import numpy as np
+
 from aizynthfinder.search.retrostar.search_tree import SearchTree
 from aizynthfinder.chem.serialization import MoleculeSerializer
 
@@ -112,10 +114,11 @@ def test_split_andor_tree(shared_datadir, default_config):
 
     routes = tree.routes()
 
-    assert len(routes) == 3
+    assert len(routes) == 97
 
 
 def test_update(shared_datadir, default_config, setup_stock):
+    # Todo: re-write
     setup_stock(
         default_config,
         "Nc1ccc(NC(=S)Nc2ccccc2)cc1",
@@ -132,8 +135,10 @@ def test_update(shared_datadir, default_config, setup_stock):
     saved_root_value = tree.root.value
     tree.mol_nodes[-1].parent.update(35, from_mol=tree.mol_nodes[-1].mol)
 
-    assert [child.value for child in tree.root.children] == [5.0, 45.0]
-    assert tree.root.value != saved_root_value
-    assert tree.root.value == 5
+    assert [np.round(child.value, 2) for child in tree.root.children][:2] == [
+        3.17,
+        3.31,
+    ]
+    assert tree.root.value == saved_root_value
 
     tree.serialize("temp.json")